+++ SayTheirNames/Source/Components/CustomSearchBar.swift
@@ -168,13 +168,3 @@ extension CustomSearchBar: UITableViewDataSource, UITableViewDelegate, UISearchB
     func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
     }
 }
-
-extension UIEdgeInsets {
-    init(top: CGFloat = 0, left: CGFloat = 0, bottom: CGFloat = 0, right: CGFloat = 0) {
-        self.init()
-        self.top = top
-        self.left = left
-        self.bottom = bottom
-        self.right = right
-    }
-}
